requirements.txt
jupyter
datasets
matplotlib
torch==1.10.2+cu113
-f https://download.pytorch.org/whl/torch_stable.html
torchvision==0.11.3+cu113
-f https://download.pytorch.org/whl/torch_stable.html
torchaudio===0.10.2+cu113
-f https://download.pytorch.org/whl/torch_stable.html
farm-haystack
transformers
pytorch_lightning
python_Levenshtein==0.12.0
spacy
sentencepiece





